requirements.txt
# ==============================================================
# POW Agent - Production Dependencies
# Complete daemon dependencies for Render deployment
# ==============================================================

# Core - Required for daemon operation
requests>=2.31.0
python-dotenv>=1.0.0
rich>=13.7.0
tenacity>=8.2.3

# Async support
aiohttp>=3.9.0
asyncio-throttle>=1.0.2

# Logging
structlog>=24.1.0

# Data validation
pydantic>=2.5.0

# CLI utilities (for setup scripts)
click>=8.1.0

# API Server (optional - for dashboard backend)
fastapi>=0.109.0
uvicorn>=0.27.0
websockets>=12.0

# Testing (dev only - not needed on Render)
# pytest>=7.4.0
# pytest-asyncio>=0.23.0
# pytest-cov>=4.1.0
# pytest-mock>=3.12.0

